GEODE-6945: expand the type to related xml attributes
